Create getmail cronjobs only if the getmail user exists.
git-svn-id: svn://svn.ispconfig.org/ispconfig3/trunk@559 c1a6f2c7-8504-0410-933a-a4dae0184b0a

+++ install/lib/installer_base.lib.php
@@ -918,25 +918,27 @@ class installer_base {
 		unlink('crontab.txt');
 		
 		//* Getmail crontab
-        $cf = $conf['getmail'];
-		exec('crontab -u getmail -l > crontab.txt');
-		$existing_cron_jobs = file('crontab.txt');
+		if(is_user('getmail')) {
+        	$cf = $conf['getmail'];
+			exec('crontab -u getmail -l > crontab.txt');
+			$existing_cron_jobs = file('crontab.txt');
 		
-		$cron_jobs = array('*/5 * * * * '.$cf['program'].' -g '.$cf['config_dir'].' -r '.$cf['config_dir'].'/*.conf > /dev/null');
+			$cron_jobs = array('*/5 * * * * '.$cf['program'].' -g '.$cf['config_dir'].' -r '.$cf['config_dir'].'/*.conf > /dev/null');
 		
-		// remove existing ispconfig cronjobs, in case the syntax has changed
-		foreach($cron_jobs as $key => $val) {
-			if(stristr($val,$cf['program'])) unset($cron_jobs[$key]);
-		}
-		
-		foreach($cron_jobs as $cron_job) {
-			if(!in_array($cron_job."\n", $existing_cron_jobs)) {
-				$existing_cron_jobs[] = $cron_job."\n";
+			// remove existing ispconfig cronjobs, in case the syntax has changed
+			foreach($cron_jobs as $key => $val) {
+				if(stristr($val,$cf['program'])) unset($cron_jobs[$key]);
 			}
+		
+			foreach($cron_jobs as $cron_job) {
+				if(!in_array($cron_job."\n", $existing_cron_jobs)) {
+					$existing_cron_jobs[] = $cron_job."\n";
+				}
+			}
+			file_put_contents('crontab.txt', $existing_cron_jobs);
+			exec('crontab -u getmail crontab.txt &> /dev/null');
+			unlink('crontab.txt');
 		}
-		file_put_contents('crontab.txt', $existing_cron_jobs);
-		exec('crontab -u getmail crontab.txt &> /dev/null');
-		unlink('crontab.txt');
 	}
 	
 }
